Try to fix multithread probs

+++ lib/common/process.rb
@@ -6,7 +6,7 @@ module Bake
 
     def self.run(cmdLineArray, immediateOutput=false, force=true, outpipe=nil, exitCodeArray = [0], dir = Dir.pwd)
       rd, wr = IO.pipe
-      @@rd = rd if force
+      @@rd = force ? rd : nil
       duppedCmdLineArray = cmdLineArray.dup
       duppedCmdLineArray << { :chdir=>dir, :err=>wr, :out=>(outpipe ? outpipe : wr) }
       begin
@@ -14,7 +14,7 @@ module Bake
       rescue Exception => e
         return [false, e.message]
       end
-      @@pid = pid if force
+      @@pid = force ? pid : nil
       wr.close
       output = ""
       begin
@@ -42,22 +42,38 @@ module Bake
       pid, status = Process.wait2(pid)
       @@pid = nil
       @@rd = nil
+
+      if status.nil? && output.empty?
+        puts output.inspect
+        output = "Process returned with unknown exit status"
+        puts output if immediateOutput
+      end
       return [false, output] if status.nil?
+
       exitCodeArray = [0] if exitCodeArray.empty?
-      [(exitCodeArray.include?status.exitstatus), output]
+      ret = (exitCodeArray.include?status.exitstatus)
+      if ret == false && output.empty?
+        output = "Process returned with exit status #{status.exitstatus}"
+        puts output if immediateOutput
+      end
+      [ret, output]
     end
 
     def self.killProcess(force) # do not kill compile processes or implement rd and pid array if really needed
-      begin
-        @@rd.close
-      rescue Exception => e
+      if @@rd
+        begin
+          @@rd.close
+        rescue Exception => e
+        end
+        @@rd = nil
       end
-      begin
-        Process.kill("KILL",@@pid)
-      rescue Exception => e
+      if @@pid
+        begin
+          Process.kill("KILL",@@pid)
+        rescue Exception => e
+        end
+        @@pid = nil
       end
-      @@rd = nil
-      @@pid = nil
     end
 
   end
